+++ packages/animator/src/index.ts
@@ -1,0 +1,4 @@
+export type { Animated } from './animated';
+export type { Animation } from './animation';
+export { Animator } from './animator';
+export { TransitionType, type Transition } from './transition';