+++ packages/spritesheet/src/index.ts
@@ -1,0 +1,3 @@
+export { animatedSpriteSheet } from './animated-sprite-sheet';
+export { spriteParser } from './sprite-parser';
+export { uniformSpriteSheet } from './uniform-sprite-sheet';